requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
langchain==0.2.6
langchain-openai==0.1.13
langchain-anthropic==0.1.20
langchain-community==0.2.6
langgraph==0.1.9
faiss-cpu==1.8.0
openai==1.35.3
anthropic==0.29.0
python-dotenv==1.0.1
pydantic==2.7.4
rank-bm25==0.2.2
python-multipart==0.0.9
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
langchain-groq==0.1.6
sentence-transformers==2.7.0
langchain-huggingface==0.0.3